==> R/07-calcular-rentabilidad-ipc.R <==
## Calcula las rentabilidades usando el ipc
library(WriteXLS)
library(tidyverse)
library(lubridate)
# Cargar IPC
ipc <- readRDS(here::here('data', 'rds', 'ipc.rds'))

# Cargar valores cuota 
# En el periodo antes de 2014 uso total para Retiro y Acumulacion
valor_cuota_total <- readRDS(here::here('data', 'rds', 'valor_cuota_total.rds'))

rentabilidad_ipc_cess <- valor_cuota_total %>% 
  # pegar ipc
  left_join(ipc, by="ano_mes") %>% 
  # calcular valor cuota/ipc
  mutate(valor_cuota_ipc = valor_cuota/ipc) %>% 
  group_by(administradora, fondo) %>% 
  # Aplicar formula del indice de rentabilidad
  mutate(rentabilidad_ipc_indice = 100 * 
           (((valor_cuota_ipc/lag(valor_cuota_ipc, 36))^(1/3)) - 1)) %>%  
  mutate(rentabilidad_ipc_simple =
           100 * (valor_cuota_ipc/lag(valor_cuota_ipc, 12) - 1)) %>% 
  ungroup()

# Salvar RDS
saveRDS(rentabilidad_ipc_cess, 
        here::here('data', 'rds', 'rentabilidad_ipc_cess.rds'))

# Salida Formato

# Acumulacion
acumulacion <- rentabilidad_ipc_cess %>% 
  filter(administradora %in% c("afap_sura", "integracion",
                             "republica", "union_capital", 
                             "sistema")) %>% 
  #filter(month(ano_mes)==12) %>% 
  filter(fondo=="Acumulación") %>% 
  select(ano_mes, administradora, rentabilidad_ipc_indice, fondo) %>% 
  pivot_wider(names_from="administradora", values_from="rentabilidad_ipc_indice") %>% 
  rename(`SURA AFAP ` = afap_sura,
         `Integración`=integracion,
         `República AFAP`=republica,
         `UNION CAPITAL`=union_capital) %>% 
  select(-fondo, fondo) %>% 
  filter(year(ano_mes) > 1999)

retiro <- rentabilidad_ipc_cess %>% 
  filter(administradora %in% c("afap_sura", "integracion",
                               "republica", "union_capital",
                               "sistema")) %>% 
  #filter(month(ano_mes)==12) %>% 
  filter(fondo=="Retiro") %>% 
  select(ano_mes, administradora, rentabilidad_ipc_indice, fondo) %>% 
  pivot_wider(names_from="administradora", values_from="rentabilidad_ipc_indice") %>% 
  rename(`SURA AFAP` = afap_sura,
         `Integración`=integracion,
         `República AFAP`=republica,
         `UNION CAPITAL`=union_capital) %>% 
  select(-fondo, fondo) %>% 
  filter(year(ano_mes) > 1999)


WriteXLS(c("retiro", "acumulacion"), 
         here::here('output', 'cuadro_rentabilidad_ipc.xlsx'))


# Salvar Excel
rentabilidad_ipc_cess %>% 
  # fondo Acumulación
  filter(fondo=="Acumulación") %>%
  # Índice 36 meses
  select(ano_mes, administradora, rentabilidad_ipc_indice) %>% 
  pivot_wider(names_from=administradora, values_from=rentabilidad_ipc_indice) %>% 
  WriteXLS(here::here('output', 'rentabilidad_ipc.xlsx'))
